cpufreq: frequency regulation based on screen on/off
[PATCH 1/6] cpufreq: frequency regulation based on screen on/off events

Cap the max cpu frequency to [min .. max/2] when the screen is off,
and to [max/2 .. max] when the screen is on.

If the cpu doesn't support exactly max/2 an appropriate frequency
is chosen, as close as possible to the theoretical value.

[PATCH 2/6] cpufreq: do not forget min/max clock frequency on cpu
hotplug

Properly implement the fix to not forget the cpu frequency range on
cpuhotplug.

Link to Source:
https://github.com/arighi/linux-gt-i9100

Signed-off-by: Andrea Righi <andrea@betterlinux.com>